refactor: change Entity name from Trend to Reel
