+++ trends_domain/src/commonMain/kotlin/net/thechance/mena/trends/domain/repository/ReelRepository.kt
@@ -1,0 +1,8 @@
+package net.thechance.mena.trends.domain.repository
+
+import net.thechance.mena.trends.domain.entity.Reel
+
+interface ReelRepository {
+    suspend fun deleteReelById(id: Int)
+    suspend fun getAllReels(): List<Reel>
+}
