--- trends_domain/src/commonMain/kotlin/net/thechance/mena/trends/domain/repository/TrendRepository.kt
@@ -1,8 +1,0 @@
-package net.thechance.mena.trends.domain.repository
-
-import net.thechance.mena.trends.domain.entity.Trend
-
-interface TrendRepository {
-    suspend fun deleteTrendById(id: Int)
-    suspend fun getAllTrends(): List<Trend>
-}
